apt: don't emit packages that are blocked or supposed to be removed
reporting blocked will get the update stuck. as an example when a package needs apt-config to be updated first, even tho its blocked now it will be unblocked after an update. don't report blocked at all because we manage it ourselves (debian does too, this is completely pointless)
reporting remove can also get updates stuck. assume libmutter-15-0 now conflicts with libmutter-14-0. now we are giving that to tryToInstall in apt-cache-file and it will try to do operation on both at the same time, but in debian/control it already says that they are conflicting and one replaces the other. so now it will fail again. instead don't report at all and let apt do its own thing (which works perfectly fine) instead of interfering with it. this also goes for lets say qt5 and 6 packages during t64 transition where one gets installed, the other is removed and there are no transitionals
we want packagekit to act like apt as much as possible which eases testing and maintenance. keeping up with two different mechanisms which enforce their own rules is a lot of pain

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>