apt: upgrade instead of dist-upgrade
this one is not specific to debian, but maybe specific to us. we test our infrastructure against upgrade. meaning sometimes to go from one release to the other, it needs to do apt update && apt upgrade && apt update && apt upgrade. this is because we handle certain changes step by step where each upgrade will unlock another set of packages. in debian they assume user is on a stable repo where there are not all that many changes. but we are on trixie (testing) which has a lot of changes and at times, some may not be very graceful. we need to handle it downstream when doing releases

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>